Include version requirement when setting generated app dev dependencies.
The development dependencies that are copied from the gemspec of the
engine to the Gemfile of the generated app were copied by name only.
This prevented someone from specifying a specific version of a development
dependency.

This commit resolves this by copying both the name and the version
requirement. If no version is specified it gets tranlated to > 0